requirements.txt
torchvision~=0.11.2
numpy~=1.21.5
scipy~=1.7.3
matplotlib~=3.5.3
scikit-image~=0.19.3